[BE] refactor: 거래서버 기능 고도화

+++ CONY-BE/payment/src/test/java/com/cony/payment/domain/payment/service/PaymentServiceTest.java
@@ -46,7 +46,7 @@ class PaymentServiceTest {
     @DisplayName("결제 준비 - TID 발급 성공")
     void ready_Success() {
         // given
-        Integer amount = 10000;
+        Long amount = 10000L;
 
         // when
         KakaoPayReadyResponse response = paymentService.ready(testUser.getId(), amount);
